Move tokenizer into module
